4) Poprawa nazwy klasy IntLinkedList
