--- src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -1,54 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class IntLinkedList {
-
-    private static final int EMPTY_STACK_VALUE = -1;
-    private Node last;
-    private int i;
-
-    public void push(final int i) {
-        if (last == null) {
-            last = new Node(i);
-        }
-        else {
-            last.next = new Node(i);
-            last.next.prev = last;
-            last = last.next;
-        }
-    }
-
-    public boolean isEmpty() {
-        return last == null;
-    }
-
-    public boolean isFull() {
-        return false;
-    }
-
-    public int top() {
-        if (isEmpty()) {
-            return EMPTY_STACK_VALUE;
-        }
-
-        return last.value;
-    }
-
-    public int pop() {
-        if (isEmpty()) {
-            return EMPTY_STACK_VALUE;
-        }
-
-        int ret = last.value;
-        last = last.prev;
-        return ret;
-    }
-}
-
-class Node {
-    int value;
-    Node prev, next;
-
-    Node(final int i) {
-        value = i;
-    }
-}
